Refine Redis client usage
